# Conflicts:
#	src/main/java/com/codewranglers/workflowmanager/models/User.java
#	src/main/java/com/codewranglers/workflowmanager/models/data/UserRepository.java

+++ src/main/java/com/codewranglers/workflowmanager/models/data/PartRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PartRepository extends CrudRepository<Part, Long> {
+public interface PartRepository extends CrudRepository<Part, Integer> {
 }
